Refactor iterator adapters into their own modules.
Also add filter, improve testing situation slightly
